backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
email-validator==2.1.0
transformers==4.48.0
torch==2.6.0
sqlalchemy==2.0.36
langdetect==1.0.9
python-multipart==0.0.6
pandas==2.2.3
python-dotenv==1.0.1
numpy==1.24.3
google-generativeai==0.8.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Message Queue & Background Tasks
redis==5.0.1
rq==1.15.1
rq-dashboard==0.7.0